requirements.txt
Flask
requests
python-dotenv
psycopg2-binary
postgresql
Flask-sqlalchemy
flask_login